models

+++ TestTimeAugmentation/mainModel.py
@@ -8,6 +8,7 @@ from imutils import paths
 
 def models(listaModels,pathImg,option):
     # 1. First we create the folder where we will store the resulting images and create as many folders as we have models
+
     os.mkdir(pathImg+'/../salida')
     for model in listaModels:
         os.mkdir(pathImg+'/../salida/'+os.path.splitext(os.path.basename(model.pathPesos))[0])
@@ -44,10 +45,10 @@ def models(listaModels,pathImg,option):
                 os.remove(file)
 
     ensembleOptions.ensembleOptions(pathImg+'/../salida/', option)
-
-    for xml in os.scandir(pathImg+'/../salida/output/'):
-        shutil.copy(pathImg+'/../salida/output/'+xml.name,pathImg+'/')
-    shutil.rmtree(pathImg+'/../salida/')
+    if notebook is False:
+        for xml in os.scandir(pathImg+'/../salida/output/'):
+            shutil.copy(pathImg+'/../salida/output/'+xml.name,pathImg+'/')
+        shutil.rmtree(pathImg+'/../salida/')
 
 
 if __name__== "__main__":
@@ -55,7 +56,7 @@ if __name__== "__main__":
     ap = argparse.ArgumentParser()
     ap.add_argument("-d", "--dataset", required=True, help="path to the dataset of images")
     ap.add_argument("-o", "--option",  default='consensus', help="option to the ensemble: affirmative, consensus or unanimous")
-
+    notebook == False
     args = vars(ap.parse_args())
     pathImg= args["dataset"]
 
@@ -66,12 +67,12 @@ if __name__== "__main__":
     imgFolder = pathImg
     #the user define configurations fichs
 
-    yoloDarknet = testTimeAugmentation.DarknetYoloPred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/AlvaroPrueba1_600train_65000.weights', '../peso/vocEstomas.names','../peso/yolov3Estomas.cfg',0.7)
-    ssdResnet = testTimeAugmentation.MXnetSSD512Pred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/ssd_512_resnet50_v1_voc-9c8b225a.params', '../peso/classesMXnet.txt',0.7)
-    fasterResnet = testTimeAugmentation.MXnetFasterRCNNPred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/faster_rcnn_resnet50_v1b_voc-447328d8.params', '../peso/classesMXnet.txt',0.7)
-    yoloResnet = testTimeAugmentation.MXnetYoloPred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/yolo3_darknet53_voc-f5ece5ce.params', '../peso/classesMXnet.txt',0.7)
-    retinaResnet50 = testTimeAugmentation.RetinaNetResnet50Pred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/resnet50_coco_best_v2.1.0.h5', '../peso/coco.csv',0.7)
-    maskRcnn = testTimeAugmentation.MaskRCNNPred('/home/ancasag/Codigo/General/ensembleObjectDetection-masterGit/peso/mask_rcnn_coco.h5', '../peso/coco.names',0.7)
+    yoloDarknet = testTimeAugmentation.DarknetYoloPred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/AlvaroPrueba1_600train_65000.weights', '../peso/vocEstomas.names','../peso/yolov3Estomas.cfg',0.7)
+    ssdResnet = testTimeAugmentation.MXnetSSD512Pred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/ssd_512_resnet50_v1_voc-9c8b225a.params', '../peso/classesMXnet.txt',0.7)
+    fasterResnet = testTimeAugmentation.MXnetFasterRCNNPred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/faster_rcnn_resnet50_v1b_voc-447328d8.params', '../peso/classesMXnet.txt',0.7)
+    yoloResnet = testTimeAugmentation.MXnetYoloPred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/yolo3_darknet53_voc-f5ece5ce.params', '../peso/classesMXnet.txt',0.7)
+    retinaResnet50 = testTimeAugmentation.RetinaNetResnet50Pred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/resnet50_coco_best_v2.1.0.h5', '../peso/coco.csv',0.7)
+    maskRcnn = testTimeAugmentation.MaskRCNNPred('/home/ancasag/Codigo/General/ensembleObjectDetection/peso/mask_rcnn_coco.h5', '../peso/coco.names',0.7)
 
     listaModels = [retinaResnet50, maskRcnn,yoloResnet,yoloDarknet,fasterResnet,ssdResnet]
